First commit!
This here is a personally tested set of code that should just output random painful sounds that increments both in the for loop and in amplitude of the sine wave used to form the WAV file! I will apply sources as I get back to it, but for now must slepp.